journals and user endpoints secured

+++ src/main/java/net/engineeringdigest/journalApp/Service/JournalEntryService.java
@@ -22,11 +22,16 @@ public class JournalEntryService {
 
     @Transactional
     public void saveEntries(JournalEntry journalEntry, String userName) {
-        User user = userService.findUserByName(userName);
-        journalEntry.setDate(LocalDateTime.now());
-        JournalEntry saved = journalEntryRepository.save(journalEntry);
-        user.getJournals().add(saved);
-        userService.saveEntries(user);
+        try {
+            User user = userService.findUserByName(userName);
+            journalEntry.setDate(LocalDateTime.now());
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+            user.getJournals().add(saved);
+            userService.saveEntries(user);
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("An error occurred while saving entry.", e);
+        }
     }
     
     public void saveEntries(JournalEntry journalEntry) {
@@ -41,10 +46,20 @@ public class JournalEntryService {
         return journalEntryRepository.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName) {
-        User user = userService.findUserByName(userName);
-        user.getJournals().removeIf(x -> x.getId().equals(id));
-        userService.saveEntries(user);
-        journalEntryRepository.deleteById(id);
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName) {
+        boolean removed = false;
+        try {
+            User user = userService.findUserByName(userName);
+            removed = user.getJournals().removeIf(x -> x.getId().equals(id));
+            if (removed) {
+                userService.saveEntries(user);
+                journalEntryRepository.deleteById(id);
+            }
+        } catch (Exception e) {
+            System.out.println(e);
+            throw new RuntimeException("Error occured while deleting record id: " + id + " for username: " + userName);
+        }
+        return removed;
     }
 }
